+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2026-06-30
+
+- Added `Return to Ward 13 '26` build tag to builds.
+
 ## 2026-06-26
 
 ### Added